removed old react template, renamed new react template, removed pre-install, updated readme to reflect removed preinstall
