--- lib/templates/template-pwa-react-new/.gitignore
@@ -1,1 +1,0 @@
-node_modules
