--- lib/templates/template-pwa-react-new/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "template-pwa-react",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js",
-    "build": "webpack -p --watch"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "express": "^4.16.3",
-    "css-loader": "^0.28.11",
-    "react": "^16.3.2",
-    "react-dom": "^16.3.2",
-    "sass-loader": "^7.0.1",
-    "style-loader": "^0.21.0",
-    "webpack": "^4.8.3"
-  },
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "nodemon": "^1.17.4",
-    "webpack-cli": "^2.1.3",
-    "webpack-dev-server": "^3.1.4"
-  }
-}
