--- lib/templates/template-pwa-react-new/webpack.config.js
@@ -1,22 +1,0 @@
-const path = require('path');
-
-const config = {
-  entry: './client/index.js',
-  output: {
-    path: path.join(__dirname, 'build'),
-    filename: 'bundle.js',
-  },
-  module: {
-    rules: [
-      { test: /\.css$/, use: 'css-loader' },
-      { test: /\.js$/, use: 'babel-loader', exclude: /node_modules/ },
-      { test: /\.jsx$/, use: 'babel-loader', exclude: /node_modules/ },
-      {
-        test: /\.scss$/,
-        use: ['style-loader', 'css-loader', 'sass-loader'],
-      },
-    ],
-  },
-};
-
-module.exports = config;
