+++ lib/templates/template-pwa-react/.babelrc
@@ -1,5 +1,3 @@
 {
-  "presets": [
-    "env", "react"
-  ]
+  "presets": ["es2015", "react", "stage-2"]
 }
