+++ lib/templates/template-pwa-react/.firebaserc
