+++ lib/templates/template-pwa-react/client/App.jsx
