+++ lib/templates/template-pwa-react/client/index.js
