+++ lib/templates/template-pwa-react/firebase.json
