+++ lib/templates/template-pwa-react/package.json
@@ -1,35 +1,31 @@
 {
-  "name": "pwa-react",
+  "name": "template-pwa-react",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
-  "dependencies": {
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.4",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "body-parser": "^1.18.3",
-    "express": "^4.16.3",
-    "lighthouse": "^2.9.4",
-    "path": "^0.12.7",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0",
-    "react-scripts": "1.1.4",
-    "serviceworker-webpack-plugin": "^0.2.3",
-    "webpack": "^4.12.0",
-    "workbox-sw": "^3.2.0"
-  },
+  "main": "server.js",
   "scripts": {
-    "start": "node server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --watch"
+    "start": "node server.js",
+    "build": "webpack -p --watch"
   },
   "author": "",
   "license": "ISC",
+  "dependencies": {
+    "express": "^4.16.3",
+    "css-loader": "^0.28.11",
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "sass-loader": "^7.0.1",
+    "style-loader": "^0.21.0",
+    "webpack": "^4.8.3"
+  },
   "devDependencies": {
-    "webpack-cli": "^3.0.3",
-    "workbox-build": "^3.2.0",
-    "workbox-webpack-plugin": "^3.2.0"
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.6.1",
+    "babel-preset-react": "^6.24.1",
+    "nodemon": "^1.17.4",
+    "webpack-cli": "^2.1.3",
+    "webpack-dev-server": "^3.1.4"
   }
 }
