--- lib/templates/template-pwa-react/public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "PWA Temp",
-  "name": "React App Template",
-  "icons": [
-    {
-      "src": "fav_Icon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": ""
-    }
-  ],
-  "start_url": "./index.html",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
