+++ lib/templates/template-pwa-react/server.js
@@ -1,33 +1,12 @@
 const express = require('express');
 const path = require('path');
-const bodyParser = require('body-parser');
+
 const app = express();
 
-// Body parser middleware
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(express.static(path.join(__dirname)));
-
-// CORS middleware
-app.use((req, res, next) => {
-  // Website you wish to allow to connect
-  res.setHeader('Access-Control-Allow-Origin', '*');
-  // Request methods you wish to allow
-  res.setHeader('Access-Control-Allow-Methods', 'GET, POST, OPTIONS, PUT, PATCH, DELETE');
-  // Request headers you wish to allow
-  res.setHeader('Access-Control-Allow-Headers', 'X-Requested-With,content-type');
-  // Set to true if you need the website to include cookies in the requests sent
-  // to the API (e.g. in case you use sessions)
-  // res.setHeader('Access-Control-Allow-Credentials', true);
-  // Pass to next layer of middleware
-  next();
-});
-
-// Set port to be used by Node.js
-app.get('/', (req, res) => { res.sendFile(path.resolve(__dirname, './public/index.html')); });
+const PORT = 3000;
 
 app.use(express.static(path.join(__dirname, 'build')));
 
-app.listen(8800, () => {
-  console.log('Node app is running on port 8800');
-});
+app.get('/', (req, res) => res.sendFile('index.html'));
+
+app.listen(PORT, console.log(`listening on... ${PORT}`));
