--- lib/templates/template-pwa-react/src/index.js
@@ -1,6 +1,0 @@
-import ReactDOM from 'react-dom';
-import React from 'react';
-import './index.css';
-import App from './App';
-
-ReactDOM.render(<App />, document.getElementById("root"));
