+++ lib/templates/template-pwa-react/webpack.config.js
@@ -1,41 +1,22 @@
 const path = require('path');
-const WorkboxPlugin = require('workbox-webpack-plugin');
-const { GenerateSW } = require('workbox-webpack-plugin');
 
-const BUILD_DIR = path.resolve(__dirname, 'build');
-
-module.exports = {
-  mode: 'development',
-  entry: './src/index.js',
+const config = {
+  entry: './client/index.js',
   output: {
+    path: path.join(__dirname, 'build'),
     filename: 'bundle.js',
-    path: BUILD_DIR,
   },
   module: {
     rules: [
+      { test: /\.css$/, use: 'css-loader' },
+      { test: /\.js$/, use: 'babel-loader', exclude: /node_modules/ },
+      { test: /\.jsx$/, use: 'babel-loader', exclude: /node_modules/ },
       {
-        test: /\.css?$/,
-        use: [
-          { loader: 'style-loader' },
-          { loader: 'css-loader' },
-        ],
-      },
-      {
-        test: /\.jsx?$/,
-        loader: 'babel-loader',
+        test: /\.scss$/,
+        use: ['style-loader', 'css-loader', 'sass-loader'],
       },
     ],
   },
-  plugins: [
-    new WorkboxPlugin.GenerateSW({
-      swDest: 'sw.js',
-      clientsClaim: true,
-      skipWaiting: true,
-      runtimeCaching: [{
-        urlPattern: new RegExp('http://localhost:8800/'),
-        handler: 'staleWhileRevalidate',
-      }],
-    }),
-  ],
 };
 
+module.exports = config;
